backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
assemblyai==0.38.1
python-dotenv==1.0.1
